+++ rss-virtual-keyboard/src/layaut/footer.js
